+++ controllers/project.js
@@ -1,15 +1,20 @@
 import { Project } from "../models/project.js";
 import { projectSchema } from "../schema/user_schema.js";
-
+import { User } from "../models/user.js";
 
 
 // Get all Projects
 export const getProjects = async (req, res, next) => {
   try {
-      const allProject = await Project.find();
-      res.json(allProject);
+    //we are fetching Project that belongs to a particular user
+    const userSessionId = req.session.user.id
+    const allProject = await Project.find({ user: userSessionId });
+    if (allProject.length == 0) {
+      return res.status(404).send("No Project added");
+    }
+    res.status(200).json({ Projects: allProject });
   } catch (error) {
-      next(error);
+    return res.status(500).json({error})
   }
 };
 
@@ -31,59 +36,84 @@ export const getProject = async (req, res, next) => {
   }
 };
 
-// Post project
+
+
+// Post/Create project
 
 export const postproject = async (req, res, next) => {
-  const { error, value } = projectSchema.validate(req.body);
-  if (error) {
-      return res.status(400).send(error.details[0].message);
-  }
   try {
-      // Add Project to database
-      const addproject = await Project.create({
-          ...req.body,
-          image: req.file ? req.file.filename : undefined, // Handle the case when req.file is not present
-      });
-      // return response
-      res.status(201).json(addproject);
+    const { error, value } = projectSchema.validate({...req.body, image:req.file.filename});
+
+    if (error) {
+      return res.status(400).send(error.details[0].message);
+    }
+
+    const userSessionId = req.session.user.id;
+   
+    const user = await User.findById(userSessionId);
+    if (!user) {
+      return res.status(404).send("User not found");
+    }
+
+    const project = await Project.create({ ...value, user: userSessionId });
+
+    user.projects.push(project._id)
+
+    await user.save();
+
+    res.status(201).json({ project });
   } catch (error) {
-      next(error);
+    console.log(error);
   }
 };
-
 
 
 // Patch Project
-export const patchProject = (req, res) => {
-  const { error, value } = projectSchema.validate(req.body);
+export const patchProject = async (req, res) => {
+try {
+  const { error, value } = projectSchema.validate({...req.body, image:req.file.filename});
   if (error) {
-    return res.status(400).send(error.details[0], message);
+    return res.status(400).send(error.details[0].message);
   }
-  res.json(`Project updated with ID ${req.params.id} updated`);
+
+  const userSessionId = req.session.user.id; 
+  const user = await User.findById(userSessionId);
+  if (!user) {
+    return res.status(404).send("User not found");
+  }
+
+  const project = await Project.findByIdAndUpdate(req.params.id, value, { new: true });
+    if (!project) {
+        return res.status(404).send("Project not found");
+    }
+
+  res.status(200).json({ project });
+} catch (error) {
+  return res.status(500).json({error})
+}
 };
 
-try {
-  // Add Project to database
-  const addproject = await Project.create({
-    ...req.body,
-    image: req.file.filename,
-  });
-  // return response
-  res.json(addproject);
-} catch (error) {
-//  next(error);
-};
 
 
 
 //Delete Project
 export const deletedProject = async (req, res, next) => {
   try {
-    // Delete by Id
-    const delProject = await Project.findByIdAndDelete(req.params.id);
-    // Return response
-    res.json(delProject);
+    const userSessionId = req.session.user.id; 
+    const user = await User.findById(userSessionId);
+    if (!user) {
+      return res.status(404).send("User not found");
+    }
+
+    const project = await Project.findByIdAndDelete(req.params.id);
+      if (!project) {
+          return res.status(404).send("Project not found");
+      }
+
+      user.projects.pull(req.params.id);
+      await user.save();
+    res.status(200).json("Project deleted");
   } catch (error) {
-    next(error);
+    return res.status(500).json({error})
   }
 };
